Refactor internal tools with manual names array and schema-derived types
Major restructuring to prepare for upcoming tools redesign:

- Move internal tools to dedicated internal-tools/ directory structure
- Create manual INTERNAL_TOOL_NAMES array to preserve literal types for z.enum()
- Separate InternalToolsSchema for proper type derivation
- Eliminate ProcessedInternalTool wrapper - use InternalTool directly
- Registry enforces completeness via TypeScript Record constraint
- Schema validates with perfect literal type inference: ("search_history")[]
- Maintain InternalToolsConfig type for future schema evolution and consistency

Architecture: Implementation drives types → Schema validates → Business logic consumes
Preparation for custom tools and tools redesign branch integration.
