+++ src/core/tools/tool-manager.ts
@@ -1,9 +1,7 @@
 import { MCPManager } from '../client/manager.js';
-import {
-    InternalToolsProvider,
-    InternalToolsServices,
-    InternalToolsConfig,
-} from './internal-tools-provider.js';
+import { InternalToolsProvider } from './internal-tools/provider.js';
+import { InternalToolsServices } from './internal-tools/registry.js';
+import type { InternalToolsConfig } from '../config/schemas.js';
 import { ToolSet } from './types.js';
 import { ToolConfirmationProvider } from './confirmation/types.js';
 import { logger } from '../logger/index.js';
